-- use 10 x C_NONCE_LIFETIME_IN_SECONDS
