+++ tests/base/src/test/java/org/keycloak/tests/oid4vc/OID4VCredentialOfferAuthCodeTest.java
@@ -4,11 +4,8 @@ import java.net.URI;
 import java.util.List;
 
 import org.keycloak.TokenVerifier;
-import org.keycloak.protocol.oid4vc.model.CredentialIssuer;
 import org.keycloak.protocol.oid4vc.model.CredentialResponse;
 import org.keycloak.protocol.oid4vc.model.CredentialsOffer;
-import org.keycloak.protocol.oid4vc.model.OID4VCAuthorizationDetail;
-import org.keycloak.protocol.oid4vc.model.Proofs;
 import org.keycloak.protocol.oid4vc.model.VerifiableCredential;
 import org.keycloak.representations.JsonWebToken;
 import org.keycloak.sdjwt.IssuerSignedJWT;
@@ -22,10 +19,8 @@ import org.keycloak.util.JsonSerialization;
 import org.junit.jupiter.api.Test;
 
 import static org.keycloak.OID4VCConstants.CLAIM_NAME_VCT;
-import static org.keycloak.OID4VCConstants.OPENID_CREDENTIAL;
 import static org.keycloak.protocol.oid4vc.issuance.OID4VCIssuerEndpoint.DEFAULT_CODE_LIFESPAN_S;
 import static org.keycloak.protocol.oidc.OIDCLoginProtocol.PROMPT_VALUE_LOGIN;
-import static org.keycloak.tests.oid4vc.OID4VCProofTestUtils.jwtProofs;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
@@ -48,80 +43,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
  */
 @KeycloakIntegrationTest(config = VCTestServerConfig.class)
 public class OID4VCredentialOfferAuthCodeTest extends OID4VCIssuerTestBase {
-
-    @Test
-    public void testNoOffer_Scope() throws Exception {
-
-        var ctx = new OID4VCTestContext(client, jwtTypeCredentialScope);
-
-        // Send AuthorizationRequest
-        //
-        AuthorizationEndpointResponse authResponse = wallet
-                .authorizationRequest()
-                .scope(ctx.getScope())
-                .send(ctx.getHolder(), "password");
-        String authCode = authResponse.getCode();
-        assertNotNull(authCode, "No authCode");
-
-        // Build and send AccessTokenRequest
-        //
-        AccessTokenResponse tokenResponse = wallet.accessTokenRequest(ctx, authCode).send();
-        String accessToken = wallet.validateHolderAccessToken(ctx, tokenResponse);
-        assertNotNull(accessToken, "No accessToken");
-
-        String authorizedIdentifier = ctx.getAuthorizedCredentialIdentifier();
-        assertNotNull(authorizedIdentifier, "No authorized credential identifier");
-
-        // Send the CredentialRequest
-        //
-        CredentialResponse credResponse = wallet.credentialRequest(ctx, accessToken)
-                .credentialIdentifier(authorizedIdentifier)
-                .proofs(getJwtProofs(ctx))
-                .send().getCredentialResponse();
-
-        verifyCredentialResponse(ctx, ctx.getHolder(), credResponse);
-    }
-
-    @Test
-    public void testNoOffer_Scope_AuthDetails() throws Exception {
-
-        var ctx = new OID4VCTestContext(client, jwtTypeCredentialScope);
-
-        CredentialIssuer issuerMetadata = wallet.getIssuerMetadata(ctx);
-
-        OID4VCAuthorizationDetail authDetail = new OID4VCAuthorizationDetail();
-        authDetail.setType(OPENID_CREDENTIAL);
-        authDetail.setCredentialConfigurationId(ctx.getCredentialConfigurationId());
-        authDetail.setLocations(List.of(issuerMetadata.getCredentialIssuer()));
-
-        // Send AuthorizationRequest
-        //
-        AuthorizationEndpointResponse authResponse = wallet
-                .authorizationRequest()
-                .scope(ctx.getScope())
-                .authorizationDetails(authDetail)
-                .send(ctx.getHolder(), "password");
-        String authCode = authResponse.getCode();
-        assertNotNull(authCode, "No authCode");
-
-        // Build and send AccessTokenRequest
-        //
-        AccessTokenResponse tokenResponse = wallet.accessTokenRequest(ctx, authCode).send();
-        String accessToken = wallet.validateHolderAccessToken(ctx, tokenResponse);
-        assertNotNull(accessToken, "No accessToken");
-
-        String authorizedIdentifier = ctx.getAuthorizedCredentialIdentifier();
-        assertNotNull(authorizedIdentifier, "No authorized credential identifier");
-
-        // Send the CredentialRequest
-        //
-        CredentialResponse credResponse = wallet.credentialRequest(ctx, accessToken)
-                .credentialIdentifier(authorizedIdentifier)
-                .proofs(getJwtProofs(ctx))
-                .send().getCredentialResponse();
-
-        verifyCredentialResponse(ctx, ctx.getHolder(), credResponse);
-    }
 
     @Test
     public void testAuthCodeOffer_Anonymous() throws Exception {
@@ -157,7 +78,7 @@ public class OID4VCredentialOfferAuthCodeTest extends OID4VCIssuerTestBase {
         //
         CredentialResponse credResponse = wallet.credentialRequest(ctx, accessToken)
                 .credentialIdentifier(authorizedIdentifier)
-                .proofs(getJwtProofs(ctx))
+                .proofs(wallet.generateJwtProof(ctx, ctx.getHolder()))
                 .send().getCredentialResponse();
 
         verifyCredentialResponse(ctx, ctx.getHolder(), credResponse);
@@ -229,7 +150,7 @@ public class OID4VCredentialOfferAuthCodeTest extends OID4VCIssuerTestBase {
         //
         CredentialResponse credResponse = wallet.credentialRequest(ctx1, accessToken1)
                 .credentialIdentifier(authorizedIdentifier1)
-                .proofs(getJwtProofs(ctx1))
+                .proofs(wallet.generateJwtProof(ctx1, ctx1.getHolder()))
                 .send().getCredentialResponse();
 
         verifyCredentialResponse(ctx1, ctx1.getHolder(), credResponse);
@@ -285,7 +206,7 @@ public class OID4VCredentialOfferAuthCodeTest extends OID4VCIssuerTestBase {
         IllegalStateException error = assertThrows(IllegalStateException.class,
                 () -> wallet.credentialRequest(ctx, accessToken)
                         .credentialIdentifier(authorizedIdentifier)
-                        .proofs(getJwtProofs(ctx))
+                        .proofs(wallet.generateJwtProof(ctx, ctx.getHolder()))
                         .send().getCredentialResponse());
         assertTrue(error.getMessage().contains("Credential offer has already expired"), error.getMessage());
         timeOffSet.set(0);
@@ -325,7 +246,7 @@ public class OID4VCredentialOfferAuthCodeTest extends OID4VCIssuerTestBase {
         //
         CredentialResponse credResponse = wallet.credentialRequest(ctx, accessToken)
                 .credentialIdentifier(authorizedIdentifier)
-                .proofs(getJwtProofs(ctx))
+                .proofs(wallet.generateJwtProof(ctx, ctx.getHolder()))
                 .send().getCredentialResponse();
 
         verifyCredentialResponse(ctx, ctx.getHolder(), credResponse);
@@ -346,11 +267,5 @@ public class OID4VCredentialOfferAuthCodeTest extends OID4VCIssuerTestBase {
         assertEquals(List.of(scope), credential.getType());
         assertEquals(URI.create("did:web:test.org"), credential.getIssuer());
         assertEquals(expUser + "@email.cz", credential.getCredentialSubject().getClaims().get("email"));
-    }
-
-    private Proofs getJwtProofs(OID4VCTestContext ctx) {
-        String cNonce = oauth.oid4vc().nonceRequest().send().getNonce();
-        String issuer = wallet.getIssuerMetadata(ctx).getCredentialIssuer();
-        return jwtProofs(issuer, cNonce);
     }
 }
